1.8.9 Minecraft Block Clutch World
From the video "How to Block Clutch in Minecraft." This world contains a machine that will teach you how to block clutch in Minecraft. 
-_-_-_-_-_-_-
To install this world... Download this file and bring it to your Desktop or somewhere easy to access.
Next unzip the file and put it into your Minecraft saves by going to "\AppData\Roaming\.minecraft\saves"
To get there, go to a folder or start button in the bottom left and type in the search bar "%appdata%". From there click the ".minecraft" folder and find the "saves" folder.
And that's it. Just drop the "1_8_9 Block Clutching" folder into the saves and you're done!
-_-_-_-_-_-_-